backend/requirements.txt
django
pillow
djangorestframework
django-cors-headers
djangorestframework-simplejwt